nits

+++ src/include/dare/dare_sm.h
@@ -45,8 +45,8 @@ typedef void (*create_snapshot_cb_t)(dare_sm_t *sm, void *snapshot);
 /* Apply a snapshot to the state machine */
 typedef int (*apply_snapshot_cb_t)(dare_sm_t *sm, void *snapshot, uint32_t size);
 
-typedef void (*proxy_store_cmd_cb_t)(int clt_id,uint8_t type,size_t data_size,void* data,void *arg);
-typedef void (*proxy_do_action_cb_t)(int clt_id,uint8_t type,size_t data_size,void* data,void *arg);
+typedef void (*proxy_store_cmd_cb_t)(uint16_t clt_id,uint8_t type,size_t data_size,void* data,void *arg);
+typedef void (*proxy_do_action_cb_t)(uint16_t clt_id,uint8_t type,size_t data_size,void* data,void *arg);
 
 struct dare_sm_t {
     destroy_cb_t   destroy;
